trim years from titles, add .1 to weights

+++ code/a_star.h
@@ -13,7 +13,7 @@ class A_Star {
   public:
   static vector<pair<string, double>> shortestPath(string start, string end,
                                      DataProcessor& data);
-  static int heuristic(int start_id, int end_id, DataProcessor& data);
+  static double heuristic(int start_id, int end_id, DataProcessor& data);
   static vector<pair<string, double>> backtrack(int start, int end,
                                   std::unordered_map<int, int>& came_from,
                                   DataProcessor& data, std::unordered_map<int, double>& g_score);
